+++ cli/migrate/database/schema_dump.go
@@ -3,5 +3,5 @@ package database
 import "github.com/hasura/graphql-engine/cli/v2/internal/hasura"
 
 type SchemaDriver interface {
-	ExportSchemaDump(schemaName []string, sourceName string, sourceKind hasura.SourceKind) ([]byte, error)
+	ExportSchemaDump(includeSchemas []string, excludeSchemas []string, sourceName string, sourceKind hasura.SourceKind) ([]byte, error)
 }